requirements.txt
flask==3.0.0
scikit-learn==1.6.1
numpy==2.2.4
joblib==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask-login==0.6.3
PyJWT==2.8.0
gevent==23.9.1
gevent-websocket==0.10.1
flask-migrate==4.0.7
